requirements.txt
# ─── Google Trends ───────────────────────────────────────────
pytrends==4.9.2

# ─── YouTube Data API + Transcripts ──────────────────────────
google-api-python-client==2.118.0
youtube-transcript-api==0.6.2

# ─── RSS Feeds ────────────────────────────────────────────────
feedparser==6.0.11

# ─── Web Scraping ─────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.2.1

# ─── AI Summarization (Groq — free, no credit card) ──────────
groq>=0.9.0

# ─── HTML Report ─────────────────────────────────────────────
jinja2==3.1.3

# ─── Utilities ───────────────────────────────────────────────
requests==2.31.0
python-dotenv==1.0.1